+++ integration-tests/src/test/java/it/capabilities/confluence/TestConfluenceConditions.java
@@ -34,6 +34,7 @@ public class TestConfluenceConditions extends ConfluenceWebDriverTestBase
     public static void startConnectAddOn() throws Exception
     {
         remotePlugin = new ConnectRunner(product.getProductInstance().getBaseUrl(),"my-plugin")
+                .setAuthenticationToNone()
                 .addModules("webItems", newWebItemBean()
                         .withName(new I18nProperty("Only Betty", ONLY_BETTY_WEBITEM))
                         .withKey("only-betty")